lo que era float cambio a double, correcciones errores de tipeo, nombres de metodos viejos

+++ src/ar/edu/unq/po2/tpIntegrador/envio/IFormaDeEnvio.java
@@ -4,6 +4,6 @@ import ar.edu.unq.po2.tpIntegrador.state.Pedido;
 
 public interface IFormaDeEnvio {
 
-	public Float calcularValorDelEnvio(Pedido pedido);
+	public Double calcularValorDelEnvio(Pedido pedido);
 	
 }
